Extract receipt renderer package
