+++ packages/printer/package.json
@@ -8,7 +8,7 @@
   "files": ["src/**"],
   "dependencies": {
     "@point-of-sale/receipt-printer-encoder": "^3.0.3",
-    "mustache": "^4.2.0",
+    "@wcpos/receipt-renderer": "workspace:*",
     "p-queue": "^9.1.1"
   },
   "devDependencies": {
