+++ packages/printer/vitest.config.ts
@@ -1,8 +1,15 @@
+import { resolve } from 'node:path';
+
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  test: {
-    globals: true,
-    environment: 'jsdom',
-  },
+	resolve: {
+		alias: {
+			'@wcpos/receipt-renderer': resolve(__dirname, '../receipt-renderer/src/index.ts'),
+		},
+	},
+	test: {
+		globals: true,
+		environment: 'jsdom',
+	},
 });
