+++ tsconfig.json
@@ -11,6 +11,8 @@
 			"@wcpos/database": ["packages/database/src/index"],
 			"@wcpos/database/*": ["packages/database/src/*"],
 			"@wcpos/hooks/*": ["packages/hooks/src/*"],
+			"@wcpos/receipt-renderer": ["packages/receipt-renderer/src/index"],
+			"@wcpos/receipt-renderer/*": ["packages/receipt-renderer/src/*"],
 			"@wcpos/printer": ["packages/printer/src/index"],
 			"@wcpos/printer/*": ["packages/printer/src/*"],
 			"@wcpos/query": ["packages/query/src/index"],
